refactor(macos): wire task selection and tighten graph/task modules
Integrate TaskMultiSelectionState into Tasks UI with a Select/Done affordance,
batch bulk delete through deleteMultipleTasks for correct score compaction,
split TaskDetailPanel into policy/navigator/view, extract knowledge graph
pagination from APIClient+Settings, consolidate TasksStore incomplete offsets,
and align AgentClient admission gate coverage for Chat-first receipt paths.
