+++ desktop/macos/Desktop/Sources/Chat/AgentContextAdmissionGate.swift
@@ -4,6 +4,15 @@ import Foundation
 /// projection for one runtime client. Swift actors are reentrant while an
 /// operation awaits the Node runtime, so actor isolation alone does not keep
 /// a context writer from advancing the snapshot between admission attempts.
+///
+/// `AgentClient.Session` routes mutating and admission-adjacent bridge calls
+/// through this gate. Intentionally excluded:
+/// - Read-only journal snapshots (`listJournalTurns`, `listJournalTurnsForControl`)
+///   — they do not advance the canonical context projection and must stay
+///   callable while projection refresh work is in flight (including from
+///   paths that already hold the gate); the gate is not reentrant.
+/// - Runtime lifecycle (`start`, `stop`, `interrupt`, auth-handler wiring) —
+///   process control, not context projection admission.
 actor AgentContextAdmissionGate {
   private var held = false
   private var waiters: [CheckedContinuation<Void, Never>] = []
