+++ desktop/macos/Desktop/Sources/MainWindow/Pages/TasksPage.swift
@@ -503,8 +503,20 @@ class TasksViewModel: ObservableObject {
   var undoToastDismissTask: Task<Void, Never>?
 
   // Multi-select state
-  @Published var isMultiSelectMode = false
-  @Published var selectedTaskIds: Set<String> = []
+  @Published private(set) var multiSelection = TaskMultiSelectionState()
+
+  var isMultiSelectMode: Bool { multiSelection.isActive }
+  var selectedTaskIds: Set<String> { multiSelection.selectedIDs }
+  var visibleTaskIDsForSelection: [String] {
+    if !showCompleted && !isMultiSelectMode {
+      return TaskCategory.allCases.flatMap { getOrderedTasks(for: $0).map(\.id) }
+    }
+    return displayTasks.map(\.id)
+  }
+
+  private var allAvailableTaskIDs: Set<String> {
+    Set(store.incompleteTasks.map(\.id) + store.completedTasks.map(\.id))
+  }
 
   // MARK: - Drag-and-Drop Reordering (like Flutter)
   /// Drag state for visual feedback
@@ -1497,6 +1509,55 @@ class TasksViewModel: ObservableObject {
 
     if keyCode == 53 { return handleEscape() }
 
+    if isMultiSelectMode {
+      if modifiers == .command && keyCode == 0 {
+        mutateMultiSelection { state in
+          _ = state.handleKeyboard(.selectAll, visibleIDs: visibleTaskIDsForSelection)
+        }
+        return true
+      }
+
+      let userModifiers = modifiers.subtracting([.numericPad, .function])
+      if keyCode == 126 && userModifiers == .shift {
+        moveSelection(-1)
+        mutateMultiSelection { state in
+          _ = state.handleKeyboard(
+            .extendFocused, focusedID: keyboardSelectedTaskId, visibleIDs: visibleTaskIDsForSelection)
+        }
+        return true
+      }
+      if keyCode == 125 && userModifiers == .shift {
+        moveSelection(1)
+        mutateMultiSelection { state in
+          _ = state.handleKeyboard(
+            .extendFocused, focusedID: keyboardSelectedTaskId, visibleIDs: visibleTaskIDsForSelection)
+        }
+        return true
+      }
+      if keyCode == 126 && userModifiers.isEmpty {
+        moveSelection(-1)
+        return true
+      }
+      if keyCode == 125 && userModifiers.isEmpty {
+        moveSelection(1)
+        return true
+      }
+      if (keyCode == 49 || keyCode == 36) && userModifiers.isEmpty {
+        mutateMultiSelection { state in
+          _ = state.handleKeyboard(
+            .toggleFocused,
+            focusedID: keyboardSelectedTaskId ?? hoveredTaskId,
+            visibleIDs: visibleTaskIDsForSelection)
+        }
+        return true
+      }
+      if modifiers == .command && keyCode == 2 {
+        Task { [weak self] in await self?.deleteSelectedTasks() }
+        return true
+      }
+      return false
+    }
+
     // Cmd+N: new task (inline at top)
     if modifiers == .command && keyCode == 45 {
       isInlineCreating = true
@@ -2319,6 +2380,8 @@ class TasksViewModel: ObservableObject {
     log(
       "TasksViewModel: Categorized \(displayTasks.count) tasks - Today: \(result[.today]?.count ?? 0), Tomorrow: \(result[.tomorrow]?.count ?? 0), Later: \(result[.later]?.count ?? 0), No Deadline: \(result[.noDeadline]?.count ?? 0)"
     )
+
+    reconcileMultiSelection()
   }
 
   /// Load more filtered/search results (pagination within already-queried results)
@@ -2608,37 +2671,80 @@ class TasksViewModel: ObservableObject {
 
   // MARK: - Multi-Select
 
+  func mutateMultiSelection(_ mutation: (inout TaskMultiSelectionState) -> Void) {
+    var next = multiSelection
+    mutation(&next)
+    multiSelection = next
+  }
+
+  private func reconcileMultiSelection() {
+    guard multiSelection.isActive else { return }
+    let visibleIDs = visibleTaskIDsForSelection
+    mutateMultiSelection { state in
+      state.reconcile(visibleIDs: visibleIDs, availableTaskIDs: allAvailableTaskIDs)
+    }
+  }
+
   func toggleMultiSelectMode() {
-    isMultiSelectMode.toggle()
-    if !isMultiSelectMode {
-      selectedTaskIds.removeAll()
+    mutateMultiSelection { state in
+      if state.isActive {
+        state.exit()
+      } else {
+        state.enter()
+        state.reconcile(visibleIDs: visibleTaskIDsForSelection)
+      }
     }
   }
 
-  func toggleTaskSelection(_ task: TaskActionItem) {
-    if selectedTaskIds.contains(task.id) {
-      selectedTaskIds.remove(task.id)
-    } else {
-      selectedTaskIds.insert(task.id)
+  func toggleTaskSelection(
+    _ task: TaskActionItem,
+    modifiers: TaskMultiSelectionModifiers = .command
+  ) {
+    mutateMultiSelection { state in
+      _ = state.click(task.id, modifiers: modifiers, visibleIDs: visibleTaskIDsForSelection)
     }
   }
 
-  func selectAll() {
-    selectedTaskIds = Set(displayTasks.map { $0.id })
+  func toggleSelectAllVisibleTasks() {
+    mutateMultiSelection { state in
+      state.toggleSelectAll(visibleIDs: visibleTaskIDsForSelection)
+    }
   }
 
-  func deselectAll() {
-    selectedTaskIds.removeAll()
+  var allVisibleTasksSelected: Bool {
+    let visibleIDs = visibleTaskIDsForSelection
+    return !visibleIDs.isEmpty && visibleIDs.allSatisfy { multiSelection.selectedIDs.contains($0) }
   }
 
   func deleteSelectedTasks() async {
-    let idsToDelete = Array(selectedTaskIds)
-    for id in idsToDelete {
+    let orderedIDs = multiSelection.selectedIDs(in: visibleTaskIDsForSelection)
+    guard !orderedIDs.isEmpty else { return }
+    guard Self.confirmBulkDelete(count: orderedIDs.count) else { return }
+
+    for id in orderedIDs {
+      removeFromDisplay(id)
       chatCoordinator?.purgeState(for: id)
     }
-    await store.deleteMultipleTasks(ids: idsToDelete)
-    selectedTaskIds.removeAll()
-    isMultiSelectMode = false
+    // One batch delete compacts relevance scores highest-first after all rows
+    // are removed; per-task delete would compact against stale score ordering.
+    await store.deleteMultipleTasks(ids: orderedIDs)
+
+    mutateMultiSelection { state in
+      state.removeSelectedIDs(Set(orderedIDs))
+      if state.selectionCount == 0 {
+        state.exit()
+      }
+    }
+  }
+
+  private static func confirmBulkDelete(count: Int) -> Bool {
+    let alert = NSAlert()
+    alert.messageText = "Delete \(count) tasks?"
+    alert.informativeText = "This cannot be undone."
+    alert.alertStyle = .warning
+    alert.addButton(withTitle: "Delete")
+    alert.addButton(withTitle: "Cancel")
+    return alert.runModal() == .alertFirstButtonReturn
   }
 
   func createTask(description: String, dueAt: Date?, priority: String?, tags: [String]? = nil) async {
@@ -3682,8 +3788,10 @@ struct TasksPage: View {
         multiSelectControls
       }
 
+      selectModeButton
+
       if viewModel.isMultiSelectMode {
-        if !viewModel.selectedTaskIds.isEmpty {
+        if viewModel.multiSelection.selectionCount > 0 {
           deleteSelectedButton
         }
         cancelMultiSelectButton
@@ -3824,29 +3932,50 @@ struct TasksPage: View {
     .help(viewModel.showCompleted ? "Hide completed tasks" : "Show completed tasks")
   }
 
+  private var selectModeButton: some View {
+    Button {
+      OmiMotion.withGated(.easeInOut(duration: 0.2)) {
+        viewModel.toggleMultiSelectMode()
+      }
+    } label: {
+      HStack(spacing: OmiSpacing.xs) {
+        Image(
+          systemName: viewModel.isMultiSelectMode ? "checkmark.circle" : "checkmark.circle.badge.questionmark"
+        )
+        .scaledFont(size: OmiType.caption)
+        Text(viewModel.isMultiSelectMode ? "Done" : "Select")
+          .scaledFont(size: OmiType.body, weight: .medium)
+      }
+      .foregroundColor(viewModel.isMultiSelectMode ? OmiColors.textPrimary : OmiColors.textSecondary)
+      .padding(.horizontal, OmiSpacing.md)
+      .padding(.vertical, OmiSpacing.sm)
+      .omiControlSurface(
+        fill: OmiColors.backgroundSecondary, radius: 18, stroke: OmiColors.border.opacity(0.18))
+    }
+    .buttonStyle(.plain)
+    .help(viewModel.isMultiSelectMode ? "Exit selection" : "Select tasks for bulk actions")
+    .accessibilityIdentifier("tasks-select-toggle")
+  }
+
   private var multiSelectControls: some View {
     HStack(spacing: OmiSpacing.md) {
       Button {
-        if viewModel.selectedTaskIds.count == viewModel.displayTasks.count {
-          viewModel.deselectAll()
-        } else {
-          viewModel.selectAll()
-        }
+        viewModel.toggleSelectAllVisibleTasks()
       } label: {
         HStack(spacing: OmiSpacing.xs) {
           Image(
-            systemName: viewModel.selectedTaskIds.count == viewModel.displayTasks.count
+            systemName: viewModel.allVisibleTasksSelected
               ? "checkmark.circle.fill" : "circle"
           )
           .scaledFont(size: OmiType.body)
-          Text(viewModel.selectedTaskIds.count == viewModel.displayTasks.count ? "Deselect All" : "Select All")
+          Text(viewModel.allVisibleTasksSelected ? "Deselect All" : "Select All")
             .scaledFont(size: OmiType.body, weight: .medium)
         }
         .foregroundColor(OmiColors.textSecondary)
       }
       .buttonStyle(.plain)
 
-      Text("\(viewModel.selectedTaskIds.count) selected")
+      Text("\(viewModel.multiSelection.selectionCount) selected")
         .scaledFont(size: OmiType.body)
         .foregroundColor(OmiColors.textTertiary)
     }
@@ -3861,7 +3990,7 @@ struct TasksPage: View {
       HStack(spacing: OmiSpacing.xs) {
         Image(systemName: "trash")
           .scaledFont(size: OmiType.caption)
-        Text("Delete \(viewModel.selectedTaskIds.count)")
+        Text("Delete \(viewModel.multiSelection.selectionCount)")
           .scaledFont(size: OmiType.body, weight: .medium)
       }
       .foregroundColor(.white)
@@ -4083,7 +4212,7 @@ struct TasksPage: View {
                   orderedTasks: orderedTasks,
                   isMultiSelectMode: viewModel.isMultiSelectMode,
                   indentLevelFor: { viewModel.getIndentLevel(for: $0) },
-                  isSelectedFor: { viewModel.selectedTaskIds.contains($0) },
+                  isSelectedFor: { viewModel.multiSelection.selectedIDs.contains($0) },
                   isKeyboardSelectedFor: { viewModel.keyboardSelectedTaskId == $0 },
                   onToggle: { await viewModel.toggleTask($0) },
                   onDelete: { await viewModel.deleteTaskWithUndo($0) },
@@ -4177,7 +4306,7 @@ struct TasksPage: View {
                   task: task,
                   indentLevel: viewModel.getIndentLevel(for: task.id),
                   isMultiSelectMode: viewModel.isMultiSelectMode,
-                  isSelected: viewModel.selectedTaskIds.contains(task.id),
+                  isSelected: viewModel.multiSelection.selectedIDs.contains(task.id),
                   isKeyboardSelected: viewModel.keyboardSelectedTaskId == task.id,
                   onToggle: { await viewModel.toggleTask($0) },
                   onDelete: { await viewModel.deleteTaskWithUndo($0) },
