+++ desktop/macos/Desktop/Sources/MainWindow/Tasks/TasksEscapeHandling.swift
@@ -1,5 +1,9 @@
 extension TasksViewModel {
   func handleEscape() -> Bool {
+    if multiSelection.isActive {
+      mutateMultiSelection { $0.exit() }
+      return true
+    }
     if isInlineCreating {
       isInlineCreating = false
       inlineCreateAfterTaskId = nil
